+++ src/@types/visitor.d.ts
@@ -5,21 +5,21 @@ import {EnumToken} from "../lib";
 /**
  * Declaration visitor handler
  */
-export declare type DeclarationVisitorHandler = (node: AstDeclaration) => AstDeclaration | AstDeclaration[] | null;
+export declare type DeclarationVisitorHandler = (node: AstDeclaration) => AstDeclaration | AstDeclaration[] | null | Promise<AstDeclaration | AstDeclaration[] | null>;
 /**
  * Rule visitor handler
  */
-export declare type RuleVisitorHandler = (node: AstRule) => AstRule | AstRule[] | null;
+export declare type RuleVisitorHandler = (node: AstRule) => AstRule | AstRule[] | null | Promise<AstRule | AstRule[] | null>;
 ;
 /**
  * AtRule visitor handler
  */
-export declare type AtRuleVisitorHandler = (node: AstAtRule) => AstAtRule | AstAtRule[] | null;
+export declare type AtRuleVisitorHandler = (node: AstAtRule) => AstAtRule | AstAtRule[] | null | Promise<AstAtRule | AstAtRule[] | null>;
 ;
 /**
  * Value visitor handler
  */
-export declare type ValueVisitorHandler = (node: Token) => Token | Token[] | null;
+export declare type ValueVisitorHandler = (node: Token) => Token | Token[] | null | Promise<Token | Token[] | null>;
 ;
 
 export declare interface VisitorNodeMap {